Modifications to permit compiling ICE programs against unmodified gadgetron code
